+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.5.3 - 2021-08-14
+
 ### Changed
 
 - Post-processors don't run if Laravel is compiling views.
@@ -9,6 +11,10 @@
 ### Added
 
 - You can set `tab_width` to `false` to output literal tabs into the rendered HTML.
+
+### Fixed
+
+- Livewire middleware won't be registered for V1 of Livewire, since it's not possible.
 
 ## 0.5.2 - 2021-08-02